requirements.txt
paramiko
pytest
pytest-mock
numpy
scikit-learn
PyYAML